backend/backend/requirements.txt
fastapi>=0.100.0
pydantic>=2.1.1
uvicorn>=0.22.0
PyJWT>=2.8.0
pytest>=7.4.0
httpx>=0.24.0
